% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{fetch_models}
\alias{fetch_models}
\title{Fetch available models from Ollama API}
\usage{
fetch_models()
}
\value{
Character vector of model names or an error message
}
\description{
Fetch available models from Ollama API
}
