% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny Application for Chat Interface}
\usage{
run_app()
}
\description{
Launches a Shiny app for interacting with the Ollama API to perform text-based chat.
The app includes options for selecting models, sending messages, viewing chat history,
and downloading chat history in HTML or CSV format.
}
\details{
The app provides a user-friendly interface with the following features:
- Model selection from available options fetched via the API.
- A text input for sending messages and receiving responses.
- A display area for chat history styled with GitHub Markdown CSS.
- Options to download the chat history in either HTML or CSV formats.
}
\examples{
if (interactive()) {
  run_app()
}
}
