% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run Ollama Chat Shiny App}
\usage{
run_app()
}
\value{
A Shiny app object.
}
\description{
This Shiny application provides an interface for chatting with an assistant using the Ollama API.
It allows users to select a model, send a message to the assistant, and view the conversation history.
The app also supports downloading the chat history in HTML or CSV formats.
}
\examples{
if (interactive()) {
  run_app()
}
}
