% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run shiny Application for Chat Interface}
\usage{
run_app()
}
\value{
No return value, called for side effects.
}
\description{
Launches a shiny app for interacting with the Ollama API to perform
text-based chat.
}
\details{
The app provides a user-friendly interface for chatting with LLMs using
Ollama, including model selection, message sending, and chat history export.
}
\examples{
if (interactive()) {
  run_app()
}
}
