% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run shiny Application for Chat Interface}
\usage{
run_app()
}
\value{
No return value, called for side effects.
}
\description{
Launches a shiny app for interacting with the Ollama API
}
